+++ src/useDebounce.ts
@@ -29,7 +29,7 @@ export default function useDebounce<T>(
   useEffect(() => {
     // We need to use this condition otherwise we will run debounce timer for the first render (including maxWait option)
     if (!eq(previousValue.current, value)) {
-      debounced.callback(value);
+      debounced(value);
       previousValue.current = value;
     }
   }, [value, debounced, eq]);